Move the main fileystemdiscover logic to utils.
So now the management command is just responsible for printing things
out to the command line.
